FU-109 feat: 예약신청서 기본정보 조회를 위한 Controller API 추가
